requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
openai==0.28.0
python-magic==0.4.27
aiofiles==23.2.1
gtts==2.3.2
python-pptx==0.6.21
moviepy==1.0.3
pillow==10.0.1
ffmpeg-python==0.2.0
asyncio==3.4.3
